+++ src/ModularPipelines.Build/Modules/PrintGitInformationModule.cs
@@ -1,0 +1,25 @@
+﻿using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.Extensions.Logging;
+using ModularPipelines.Context;
+using ModularPipelines.Git.Extensions;
+using ModularPipelines.Modules;
+
+namespace ModularPipelines.Build.Modules;
+
+public class PrintGitInformationModule : Module
+{
+    protected override async Task<IDictionary<string, object>?> ExecuteAsync(IPipelineContext context, CancellationToken cancellationToken)
+    {
+        await Task.CompletedTask;
+        
+        context.Logger.LogInformation("Git Info: {GitInfo}", JsonSerializer.Serialize(context.Git().Information, new JsonSerializerOptions
+        {
+            ReferenceHandler = ReferenceHandler.IgnoreCycles,
+            IgnoreReadOnlyFields = true,
+            WriteIndented = true,
+        }));
+        
+        return null;
+    }
+}