+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
-*   DependsOnAllModulesInheritingFromAttribute to allow a module to wait on all modules inheriting from a base without needing to list each module explicitly
+*   `[DependsOnAllModulesInheritingFromAttribute]` to allow a module to wait on all modules inheriting from a base without needing to list each module explicitly
+*   Better JSON serialization of `Folder` objects